refactor(simulator): move simulator helpers out of guest-agent crate
